+++ examples/echo_server.rs
@@ -1,4 +1,5 @@
-use reactio::{example, logmsg, DefaultTcpListenerHandler, ReactRuntime};
+use reactio::{logmsg, DefaultTcpListenerHandler, ReactRuntime};
+mod example;
 
 fn run(port: i32) {
     let addr = "127.0.0.1:".to_owned() + &port.to_string();
@@ -8,9 +9,9 @@ fn run(port: i32) {
     cmd_sender
         .send_listen(
             &addr,
-            DefaultTcpListenerHandler::<example::MyReactor>::new(
+            DefaultTcpListenerHandler::<example::pingpong::PingpongReactor>::new(
                 recv_buffer_min_size,
-                example::ServerParam {
+                example::pingpong::ServerParam {
                     name: "server".to_owned(),
                     latency_batch: 1000,
                 },
@@ -26,6 +27,7 @@ fn run(port: i32) {
     while runtime.process_events() {}
     assert_eq!(runtime.count_reactors(), 0);
 }
+
 fn main() {
     let port = 10254;
     run(port);